Reports plugin complete
